Rename test packages 'samples' -> 'specimen'
